# Optimize tracing terminal in sys.