<!-- Please check for duplicates or similar PRs before submitting this PR. -->
## Does this PR close any issues? If so, link them below.
N/A

## Briefly describe the issue(s) fixed.
Makes tracing terminal faster by removing some of the freezes.