@lemz1 2024-09-21
wait, how does this make it faster?
wouldn't using mutexes, make it technically slower, since you lock the threads, until the mutex is released?

also isn't printing done synchronously, meaning on a singular thread?