@EliteMasterEric 2024-09-25
I now have the tools to get objective data for this!

I performed the following test code, then filtered the statistics such that I was timing just the `trace()` call.

```haxe
  function debug_testTraces():Void
  {
    for (i in 0...10000)
    {
      trace('test $i');
    }
  }
```

**Results before this PR:**

![image](https://github.com/user-attachments/assets/756548ad-0cf3-4ccc-b087-6a2100be9f47)

We see a cumulative time of 423.6ms with a mean of 42.36us and a max of 5.22ms.

**Results after this PR:**

![image](https://github.com/user-attachments/assets/7cebbb82-7395-498b-bd72-9dd4bc1b5fb1)

We see a cumulative time of 478.18ms with a mean of 47.82us and a max of 4.14ms.

We also have this time distribution for `debug_testTraces` as a whole which I forgot to grab for the first test.

![image](https://github.com/user-attachments/assets/0778e800-a396-42cd-9572-1c503f8a88d0)

## Conclusion

Reading the code itself, I'm not 100% sure what wrapping the trace call in a mutex would do here. Maybe you misunderstood how multi-threading works? Locking a mutex does NOT create a new thread to do work in, it's for managing work done by threads!

I'm not 100% certain what makes each individual call to `trace()` take the time it does (probably something related to the lower level system) but overall I found this PR had a slightly negative but overall negligible impact on performance, rather than the positive one this claims to have.

Feel free to reply if you have any objections but for now I will reject this PR.